+++ x/auth/vesting/types/vesting_account_test.go
@@ -16,6 +16,7 @@ import (
 	"cosmossdk.io/x/auth/vesting"
 	"cosmossdk.io/x/auth/vesting/types"
 
+	codectestutil "github.com/cosmos/cosmos-sdk/codec/testutil"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
 	"github.com/cosmos/cosmos-sdk/runtime"
 	"github.com/cosmos/cosmos-sdk/testutil"
@@ -38,7 +39,7 @@ type VestingAccountTestSuite struct {
 }
 
 func (s *VestingAccountTestSuite) SetupTest() {
-	encCfg := moduletestutil.MakeTestEncodingConfig(vesting.AppModule{})
+	encCfg := moduletestutil.MakeTestEncodingConfig(codectestutil.CodecOptions{}, vesting.AppModule{})
 
 	key := storetypes.NewKVStoreKey(authtypes.StoreKey)
 	env := runtime.NewEnvironment(runtime.NewKVStoreService(key), log.NewNopLogger())